refactor: Change folder structure.
