+++ lua/diffview/api/views/diff/diff_view.lua
@@ -1,8 +1,8 @@
 local CFileEntry = require("diffview.api.views.file_entry").CFileEntry
-local DiffView = require("diffview.views.diff.diff_view").DiffView
+local DiffView = require("diffview.scene.views.diff.diff_view").DiffView
 local EventEmitter = require("diffview.events").EventEmitter
 local FileDict = require("diffview.git.file_dict").FileDict
-local FilePanel = require("diffview.views.diff.file_panel").FilePanel
+local FilePanel = require("diffview.scene.views.diff.file_panel").FilePanel
 local Rev = require("diffview.git.rev").Rev
 local RevType = require("diffview.git.rev").RevType
 local async = require("plenary.async")
