+++ lua/diffview/lib.lua
@@ -1,15 +1,15 @@
 local lazy = require("diffview.lazy")
 
 ---@type DiffView|LazyModule
-local DiffView = lazy.access("diffview.views.diff.diff_view", "DiffView")
+local DiffView = lazy.access("diffview.scene.views.diff.diff_view", "DiffView")
 ---@type FileHistoryView|LazyModule
-local FileHistoryView = lazy.access("diffview.views.file_history.file_history_view", "FileHistoryView")
+local FileHistoryView = lazy.access("diffview.scene.views.file_history.file_history_view", "FileHistoryView")
 ---@type Rev|LazyModule
 local Rev = lazy.access("diffview.git.rev", "Rev")
 ---@type RevType|LazyModule
 local RevType = lazy.access("diffview.git.rev", "RevType")
 ---@type StandardView|LazyModule
-local StandardView = lazy.access("diffview.views.standard.standard_view", "StandardView")
+local StandardView = lazy.access("diffview.scene.views.standard.standard_view", "StandardView")
 ---@module "diffview.arg_parser"
 local arg_parser = lazy.require("diffview.arg_parser")
 ---@module "diffview.config"
