+++ lua/diffview/scene/file_entry.lua
