+++ lua/diffview/scene/views/diff/render.lua
